Add name resolution step
Also refactor error handling to be more in line with the book.
Also massive refactor to use module imports and fully qualified names in a lot
of places, which was needed to deal with the tight coupling of so many modules
and issues with circular imports and mypy being weird. There must be a better
way to do it but this will do for now.